move non-windows into a new file as well.
