+++ codex-rs/core/src/command_safety/mod.rs
@@ -1,3 +1,5 @@
 pub mod is_safe_command;
+#[cfg(unix)]
+pub mod unix_safe_commands;
 #[cfg(target_os = "windows")]
 pub mod windows_safe_commands;
